+++ CHANGELOG.md
@@ -5,11 +5,18 @@ CHANGELOG
 dev
 ---
 
+ ### New features
+
+ * ons-toolbar: Added new `static` attribute to avoid animations in the toolbar when pushing or popping pages.
+
  ### Bug Fixes
- 
+
  * ons-tabbar: Fixed [#2316](https://github.com/OnsenUI/OnsenUI/issues/2316).
+ * ons-tabbar: Fixed [#2343](https://github.com/OnsenUI/OnsenUI/issues/2343).
  * ons-splitter-side: Fixed [#2271](https://github.com/OnsenUI/OnsenUI/issues/2271).
  * ons-navigator: Fixed [#2333](https://github.com/OnsenUI/OnsenUI/issues/2333).
+ * ons-select (fastclick): Fixed [#2352](https://github.com/OnsenUI/OnsenUI/issues/2352).
+ * ons-switch: Fixed [#2341](https://github.com/OnsenUI/OnsenUI/issues/2341).
  
  ### Misc